move some more relation-related functions to separate file
