+++ packages/nocodb/src/db/BaseModelSqlv2/relation-list-count.ts
@@ -1,12 +1,13 @@
 import { extractFilterFromXwhere, NcApiVersion } from 'nocodb-sdk';
 import groupBy from 'lodash/groupBy';
 import type { Logger } from '@nestjs/common';
-import type { IBaseModelSqlV2 } from '../IBaseModelSqlV2';
+import type { IBaseModelSqlV2 } from '~/db/IBaseModelSqlV2';
 import type { LinkToAnotherRecordColumn } from '~/models';
-import { _wherePk } from '~/helpers/dbHelpers';
+import { _wherePk, applyPaginate } from '~/helpers/dbHelpers';
 import { Filter, Model, View } from '~/models';
 import sortV2 from '~/db/sortV2';
 import conditionV2 from '~/db/conditionV2';
+import getAst from '~/helpers/getAst';
 
 const GROUP_COL = '__nc_group_id';
 
@@ -717,6 +718,766 @@ export const relationListCount = (param: {
       );
       return (await this.execAndParse(qb, null, { raw: true, first: true }))
         ?.count;
+    },
+
+    async getMmChildrenExcludedListCount(
+      { colId, pid = null },
+      args,
+    ): Promise<any> {
+      const { where } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const mmTable = await relColOptions.getMMModel(this.context);
+      const assocBaseModel = await Model.getBaseModelSQL(this.context, {
+        id: mmTable.id,
+        dbDriver: this.dbDriver,
+      });
+
+      const vtn = assocBaseModel.getTnPath(mmTable);
+      const vcn = (await relColOptions.getMMChildColumn(this.context))
+        .column_name;
+      const vrcn = (await relColOptions.getMMParentColumn(this.context))
+        .column_name;
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const childTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+
+      const childView = await relColOptions.getChildView(this.context);
+      let listArgs: any = {};
+      if (childView) {
+        const { dependencyFields } = await getAst(this.context, {
+          model: childTable,
+          query: {},
+          view: childView,
+          throwErrorIfInvalidParams: false,
+        });
+
+        listArgs = dependencyFields;
+        try {
+          listArgs.filterArr = JSON.parse(listArgs.filterArrJson);
+        } catch (e) {}
+        try {
+          listArgs.sortArr = JSON.parse(listArgs.sortArrJson);
+        } catch (e) {}
+      }
+
+      const parentTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+      await parentTable.getColumns(this.context);
+
+      const parentBaseModel = await Model.getBaseModelSQL(this.context, {
+        id: parentTable.id,
+        dbDriver: this.dbDriver,
+      });
+      const childBaseModel = await Model.getBaseModelSQL(this.context, {
+        id: childTable.id,
+        dbDriver: this.dbDriver,
+      });
+      const childTn = childBaseModel.getTnPath(childTable);
+      const parentTn = parentBaseModel.getTnPath(parentTable);
+
+      const rtn = childTn;
+      const qb = this.dbDriver(rtn)
+        .count(`*`, { as: 'count' })
+        .where((qb) => {
+          qb.whereNotIn(
+            rcn,
+            this.dbDriver(rtn)
+              .select(`${rtn}.${rcn}`)
+              .join(vtn, `${rtn}.${rcn}`, `${vtn}.${vrcn}`)
+              .whereIn(
+                `${vtn}.${vcn}`,
+                this.dbDriver(parentTn)
+                  .select(cn)
+                  // .where(parentTable.primaryKey.cn, pid)
+                  .where(_wherePk(parentTable.primaryKeys, pid)),
+              ),
+          ).orWhereNull(rcn);
+        });
+
+      const aliasColObjMap = await childTable.getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: childView,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: pid,
+      });
+
+      return (
+        await this.execAndParse(qb, await childTable.getColumns(this.context), {
+          raw: true,
+          first: true,
+        })
+      )?.count;
+    },
+
+    async getMmChildrenExcludedList({ colId, pid = null }, args): Promise<any> {
+      const { where, sort, ...rest } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const mmTable = await relColOptions.getMMModel(this.context);
+      const assocBaseModel = await Model.getBaseModelSQL(this.context, {
+        id: mmTable.id,
+        dbDriver: this.dbDriver,
+      });
+
+      const vtn = assocBaseModel.getTnPath(mmTable);
+      const vcn = (await relColOptions.getMMChildColumn(this.context))
+        .column_name;
+      const vrcn = (await relColOptions.getMMParentColumn(this.context))
+        .column_name;
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+
+      const childTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+      const parentTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+      await parentTable.getColumns(this.context);
+      const parentBaseModel = await Model.getBaseModelSQL(this.context, {
+        id: parentTable.id,
+        dbDriver: this.dbDriver,
+      });
+      const childBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        id: childTable.id,
+      });
+      const childTn = childBaseModel.getTnPath(childTable);
+      const parentTn = parentBaseModel.getTnPath(parentTable);
+
+      const childView = await relColOptions.getChildView(
+        this.context,
+        childTable,
+      );
+      let listArgs: any = {};
+      if (childView) {
+        const { dependencyFields } = await getAst(this.context, {
+          model: childTable,
+          query: {},
+          view: childView,
+          throwErrorIfInvalidParams: false,
+        });
+        listArgs = dependencyFields;
+      }
+
+      const rtn = childTn;
+
+      const qb = this.dbDriver(rtn).where((qb) =>
+        qb
+          .whereNotIn(
+            rcn,
+            this.dbDriver(rtn)
+              .select(`${rtn}.${rcn}`)
+              .join(vtn, `${rtn}.${rcn}`, `${vtn}.${vrcn}`)
+              .whereIn(
+                `${vtn}.${vcn}`,
+                this.dbDriver(parentTn)
+                  .select(cn)
+                  // .where(parentTable.primaryKey.cn, pid)
+                  .where(_wherePk(parentTable.primaryKeys, pid)),
+              ),
+          )
+          .orWhereNull(rcn),
+      );
+
+      if (+rest?.shuffle) {
+        await this.shuffle({ qb });
+      }
+
+      await childBaseModel.selectObject({
+        qb,
+        fieldsSet: listArgs?.fieldsSet,
+        viewId: childView?.id,
+      });
+
+      const aliasColObjMap = await childTable.getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: relColOptions.fk_target_view_id ? childView : null,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: pid,
+      });
+
+      await this.applySortAndFilter({
+        table: childTable,
+        view: childView,
+        qb,
+        sort,
+        where,
+        // condition is applied in getCustomConditionsAndApply and we don't want to apply it again
+        onlySort: true,
+      });
+
+      applyPaginate(qb, rest);
+
+      const proto = await childBaseModel.getProto();
+      const data = await this.execAndParse(
+        qb,
+        await childTable.getColumns(this.context),
+      );
+      return data.map((c) => {
+        c.__proto__ = proto;
+        return c;
+      });
+    },
+
+    async getHmChildrenExcludedList({ colId, pid = null }, args): Promise<any> {
+      const { where, sort, ...rest } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const childTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+      const parentTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+      const childBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: childTable,
+      });
+      const parentBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: parentTable,
+      });
+      await parentTable.getColumns(this.context);
+
+      const childView = await relColOptions.getChildView(
+        this.context,
+        childTable,
+      );
+
+      const childTn = childBaseModel.getTnPath(childTable);
+      const parentTn = parentBaseModel.getTnPath(parentTable);
+
+      const tn = childTn;
+      const rtn = parentTn;
+
+      const qb = this.dbDriver(tn).where((qb) => {
+        qb.whereNotIn(
+          cn,
+          this.dbDriver(rtn)
+            .select(rcn)
+            // .where(parentTable.primaryKey.cn, pid)
+            .where(_wherePk(parentTable.primaryKeys, pid)),
+        ).orWhereNull(cn);
+      });
+
+      if (+rest?.shuffle) {
+        await this.shuffle({ qb });
+      }
+
+      await childBaseModel.selectObject({ qb });
+
+      const aliasColObjMap = await childTable.getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: relColOptions.fk_target_view_id ? childView : null,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: pid,
+      });
+      await this.applySortAndFilter({
+        table: childTable,
+        view: childView,
+        qb,
+        sort,
+        where,
+        // condition is applied in getCustomConditionsAndApply and we don't want to apply it again
+        onlySort: true,
+      });
+
+      applyPaginate(qb, rest);
+
+      const proto = await childBaseModel.getProto();
+      const data = await this.execAndParse(
+        qb,
+        await childTable.getColumns(this.context),
+      );
+      return data.map((c) => {
+        c.__proto__ = proto;
+        return c;
+      });
+    },
+
+    async getHmChildrenExcludedListCount(
+      { colId, pid = null },
+      args,
+    ): Promise<any> {
+      const { where } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const childTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+      const parentTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+
+      const childView = await relColOptions.getChildView(this.context);
+
+      const childBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: childTable,
+      });
+
+      const childTn = childBaseModel.getTnPath(childTable);
+      const parentTn = this.getTnPath(parentTable);
+
+      const tn = childTn;
+      const rtn = parentTn;
+      await parentTable.getColumns(this.context);
+
+      const qb = this.dbDriver(tn)
+        .count(`*`, { as: 'count' })
+        .where((qb) => {
+          qb.whereNotIn(
+            cn,
+            this.dbDriver(rtn)
+              .select(rcn)
+              // .where(parentTable.primaryKey.cn, pid)
+              .where(_wherePk(parentTable.primaryKeys, pid)),
+          ).orWhereNull(cn);
+        });
+
+      const aliasColObjMap = await childTable.getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: childView,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: pid,
+      });
+
+      return (await this.execAndParse(qb, null, { raw: true, first: true }))
+        ?.count;
+    },
+
+    async getExcludedOneToOneChildrenList(
+      { colId, cid = null },
+      args,
+    ): Promise<any> {
+      const { where, sort, ...rest } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const parentTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const childTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+      const parentModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: parentTable,
+      });
+      const childModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: childTable,
+      });
+
+      // one-to-one relation is combination of both hm and bt to identify table which have
+      // foreign key column(similar to bt) we are adding a boolean flag `bt` under meta
+      const isBt = relColumn.meta?.bt;
+
+      const targetView = await relColOptions.getChildView(
+        this.context,
+        isBt ? parentTable : childTable,
+      );
+      let listArgs: any = {};
+      if (targetView) {
+        const { dependencyFields } = await getAst(this.context, {
+          model: isBt ? parentTable : childTable,
+          query: {},
+          view: targetView,
+          throwErrorIfInvalidParams: false,
+        });
+        listArgs = dependencyFields;
+      }
+
+      const rtn = this.getTnPath(parentTable);
+      const tn = this.getTnPath(childTable);
+      await childTable.getColumns(this.context);
+
+      const qb = this.dbDriver(isBt ? rtn : tn).where((qb) => {
+        qb.whereNotIn(
+          isBt ? rcn : cn,
+          this.dbDriver(isBt ? tn : rtn)
+            .select(isBt ? cn : rcn)
+            .where(_wherePk((isBt ? childTable : parentTable).primaryKeys, cid))
+            .whereNotNull(isBt ? cn : rcn),
+        ).orWhereNull(isBt ? rcn : cn);
+      });
+
+      if (+rest?.shuffle) {
+        await this.shuffle({ qb });
+      }
+
+      // pre-load columns for later user
+      await parentTable.getColumns(this.context);
+      await childTable.getColumns(this.context);
+
+      await (isBt ? parentModel : childModel).selectObject({
+        qb,
+        fieldsSet: listArgs.fieldsSet,
+        viewId: targetView?.id,
+      });
+
+      // extract col-alias map based on the correct relation table
+      const aliasColObjMap = await (relColumn.meta?.bt
+        ? parentTable
+        : childTable
+      ).getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: relColOptions.fk_target_view_id ? targetView : null,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: cid,
+      });
+
+      await this.applySortAndFilter({
+        table: isBt ? parentTable : childTable,
+        view: targetView,
+        qb,
+        sort,
+        where,
+        // condition is applied in getCustomConditionsAndApply and we don't want to apply it again
+        onlySort: true,
+      });
+
+      applyPaginate(qb, rest);
+
+      const proto = await (isBt ? parentModel : childModel).getProto();
+      const data = await this.execAndParse(
+        qb,
+        await (isBt ? parentTable : childTable).getColumns(this.context),
+      );
+
+      return data.map((c) => {
+        c.__proto__ = proto;
+        return c;
+      });
+    },
+
+    async getBtChildrenExcludedListCount(
+      { colId, cid = null },
+      args,
+    ): Promise<any> {
+      const { where } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const parentTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const childTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+
+      const parentBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: parentTable,
+      });
+
+      const childTn = this.getTnPath(childTable);
+      const parentTn = parentBaseModel.getTnPath(parentTable);
+
+      const rtn = parentTn;
+      const tn = childTn;
+      await childTable.getColumns(this.context);
+
+      const qb = this.dbDriver(rtn)
+        .where((qb) => {
+          qb.whereNotIn(
+            rcn,
+            this.dbDriver(tn)
+              .select(cn)
+              // .where(childTable.primaryKey.cn, cid)
+              .where(_wherePk(childTable.primaryKeys, cid))
+              .whereNotNull(cn),
+          );
+        })
+        .count(`*`, { as: 'count' });
+
+      const aliasColObjMap = await parentTable.getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      const targetView = await relColOptions.getChildView(this.context);
+
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: targetView,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: cid,
+      });
+
+      return (await this.execAndParse(qb, null, { raw: true, first: true }))
+        ?.count;
+    },
+
+    async countExcludedOneToOneChildren(
+      { colId, cid = null },
+      args,
+    ): Promise<any> {
+      const { where } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const parentTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const childTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+
+      const childView = await relColOptions.getChildView(this.context);
+      const parentBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: parentTable,
+      });
+      const childBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: childTable,
+      });
+      const childTn = childBaseModel.getTnPath(childTable);
+      const parentTn = parentBaseModel.getTnPath(parentTable);
+
+      const rtn = parentTn;
+      const tn = childTn;
+
+      // pre-load columns for later user
+      await childTable.getColumns(this.context);
+      await parentTable.getColumns(this.context);
+
+      // one-to-one relation is combination of both hm and bt to identify table which have
+      // foreign key column(similar to bt) we are adding a boolean flag `bt` under meta
+      const isBt = relColumn.meta?.bt;
+
+      const qb = this.dbDriver(isBt ? rtn : tn)
+        .where((qb) => {
+          qb.whereNotIn(
+            isBt ? rcn : cn,
+            this.dbDriver(isBt ? tn : rtn)
+              .select(isBt ? cn : rcn)
+              .where(
+                _wherePk((isBt ? childTable : parentTable).primaryKeys, cid),
+              )
+              .whereNotNull(isBt ? cn : rcn),
+          ).orWhereNull(isBt ? rcn : cn);
+        })
+        .count(`*`, { as: 'count' });
+
+      // extract col-alias map based on the correct relation table
+      const aliasColObjMap = await (relColumn.meta?.bt
+        ? parentTable
+        : childTable
+      ).getAliasColObjMap(this.context);
+
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: childView,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: cid,
+      });
+
+      return (await this.execAndParse(qb, null, { raw: true, first: true }))
+        ?.count;
+    },
+
+    async getBtChildrenExcludedList({ colId, cid = null }, args): Promise<any> {
+      const { where, sort, ...rest } = this._getListArgs(args as any);
+      const relColumn = (await this.model.getColumns(this.context)).find(
+        (c) => c.id === colId,
+      );
+      const relColOptions = (await relColumn.getColOptions(
+        this.context,
+      )) as LinkToAnotherRecordColumn;
+
+      const rcn = (await relColOptions.getParentColumn(this.context))
+        .column_name;
+      const parentTable = await (
+        await relColOptions.getParentColumn(this.context)
+      ).getModel(this.context);
+      const cn = (await relColOptions.getChildColumn(this.context)).column_name;
+      const childTable = await (
+        await relColOptions.getChildColumn(this.context)
+      ).getModel(this.context);
+      const parentBaseModel = await Model.getBaseModelSQL(this.context, {
+        dbDriver: this.dbDriver,
+        model: parentTable,
+      });
+
+      const childTn = this.getTnPath(childTable);
+      const parentTn = parentBaseModel.getTnPath(parentTable);
+
+      const rtn = parentTn;
+      const tn = childTn;
+      await childTable.getColumns(this.context);
+
+      const qb = this.dbDriver(rtn).where((qb) => {
+        qb.whereNotIn(
+          rcn,
+          this.dbDriver(tn)
+            .select(cn)
+            // .where(childTable.primaryKey.cn, cid)
+            .where(_wherePk(childTable.primaryKeys, cid))
+            .whereNotNull(cn),
+        );
+      });
+
+      if (+rest?.shuffle) {
+        await this.shuffle({ qb });
+      }
+
+      await parentBaseModel.selectObject({ qb });
+
+      const aliasColObjMap = await parentTable.getAliasColObjMap(this.context);
+      const { filters: filterObj } = extractFilterFromXwhere(
+        this.context,
+        where,
+        aliasColObjMap,
+      );
+
+      const targetView = await relColOptions.getChildView(
+        this.context,
+        parentTable,
+      );
+      await this.getCustomConditionsAndApply({
+        column: relColumn,
+        view: relColOptions.fk_target_view_id ? targetView : null,
+        filters: filterObj,
+        args,
+        qb,
+        rowId: cid,
+      });
+
+      await this.applySortAndFilter({
+        table: parentTable,
+        view: targetView,
+        qb,
+        sort,
+        where,
+        // condition is applied in getCustomConditionsAndApply and we don't want to apply it again
+        onlySort: true,
+      });
+
+      applyPaginate(qb, rest);
+
+      const proto = await parentBaseModel.getProto();
+      const data = await this.execAndParse(
+        qb,
+        await parentTable.getColumns(this.context),
+      );
+
+      return data.map((c) => {
+        c.__proto__ = proto;
+        return c;
+      });
     },
   };
 };
