+++ packages/nocodb/src/helpers/dbHelpers.ts
@@ -2,24 +2,35 @@ import {
   isCreatedOrLastModifiedByCol,
   isCreatedOrLastModifiedTimeCol,
   isOrderCol,
+  isSystemColumn,
+  isVirtualCol,
+  NcApiVersion,
   type NcContext,
   RelationTypes,
   UITypes,
 } from 'nocodb-sdk';
 import { customAlphabet } from 'nanoid';
 import { v4 as uuidv4 } from 'uuid';
-import type { BaseModelSqlv2 } from 'src/db/BaseModelSqlv2';
-import type CustomKnex from 'src/db/CustomKnex';
+import type {
+  XcFilter,
+  XcFilterWithAlias,
+} from '~/db/sql-data-mapper/lib/BaseModel';
+import type { SortType } from 'nocodb-sdk';
+import type CustomKnex from '~/db/CustomKnex';
 import type { Knex } from 'knex';
-import type { Filter } from '~/models';
+import type { Filter, GridViewColumn } from '~/models';
+import type { BaseModelSqlv2 } from '~/db/BaseModelSqlv2';
+import { defaultLimitConfig } from '~/helpers/extractLimitAndOffset';
 import NcConnectionMgrv2 from '~/utils/common/NcConnectionMgrv2';
 import {
   Column,
   type LinkToAnotherRecordColumn,
   Model,
+  Sort,
   Source,
 } from '~/models';
 import { NcError } from '~/helpers/catchError';
+import { excludeAttachmentProps } from '~/utils';
 
 export function concatKnexRaw(knex: CustomKnex, raws: Knex.Raw[]) {
   return knex.raw(raws.map(() => '?').join(' '), raws);
@@ -307,4 +318,187 @@ export function transformObject(value, idToAliasMap) {
     }
   });
   return result;
+}
+
+export function extractSortsObject(
+  _sorts: string | string[],
+  aliasColObjMap: { [columnAlias: string]: Column },
+  throwErrorIfInvalid = false,
+): Sort[] {
+  if (!_sorts?.length) return;
+
+  let sorts = _sorts;
+  if (!Array.isArray(sorts)) sorts = sorts.split(/\s*,\s*/);
+
+  return sorts.map((s) => {
+    const sort: SortType = { direction: 'asc' };
+    if (s.startsWith('-')) {
+      sort.direction = 'desc';
+      sort.fk_column_id = aliasColObjMap[s.slice(1)]?.id;
+    } else if (s.startsWith('~-')) {
+      sort.direction = 'count-desc';
+      sort.fk_column_id = aliasColObjMap[s.slice(2)]?.id;
+    } else if (s.startsWith('~+')) {
+      sort.direction = 'count-asc';
+      sort.fk_column_id = aliasColObjMap[s.slice(2)]?.id;
+    }
+    // replace + at the beginning if present
+    else {
+      sort.fk_column_id = aliasColObjMap[s.replace(/^\+/, '')]?.id;
+    }
+
+    if (throwErrorIfInvalid && !sort.fk_column_id)
+      NcError.fieldNotFound(s.replace(/^[+-]/, ''));
+    return new Sort(sort);
+  });
+}
+
+export function applyPaginate(
+  query,
+  {
+    limit = 25,
+    offset = 0,
+    ignoreLimit = false,
+  }: XcFilter & { ignoreLimit?: boolean },
+) {
+  query.offset(offset);
+  if (!ignoreLimit) query.limit(limit);
+  return query;
+}
+
+export function haveFormulaColumn(columns: Column[]) {
+  return columns.some((c) => c.uidt === UITypes.Formula);
+}
+
+export function shouldSkipField(
+  fieldsSet,
+  viewOrTableColumn,
+  view,
+  column,
+  extractPkAndPv,
+) {
+  if (fieldsSet) {
+    return !fieldsSet.has(column.title);
+  } else {
+    if (column.system && isCreatedOrLastModifiedByCol(column)) return true;
+    if (column.system && isOrderCol(column)) return true;
+    if (!extractPkAndPv) {
+      if (!(viewOrTableColumn instanceof Column)) {
+        if (
+          !(viewOrTableColumn as GridViewColumn)?.show &&
+          !(column.rqd && !column.cdf && !column.ai) &&
+          !column.pk &&
+          column.uidt !== UITypes.ForeignKey
+        )
+          return true;
+        if (
+          !view?.show_system_fields &&
+          column.uidt !== UITypes.ForeignKey &&
+          !column.pk &&
+          isSystemColumn(column)
+        )
+          return true;
+      }
+    }
+    return false;
+  }
+}
+
+export function getListArgs(
+  args: XcFilterWithAlias,
+  model: Model,
+  {
+    ignoreAssigningWildcardSelect = false,
+    apiVersion = NcApiVersion.V2,
+    nested = false,
+  } = {},
+): XcFilter {
+  const obj: XcFilter = {};
+  obj.where = args.where || args.filter || args.w || '';
+  obj.having = args.having || args.h || '';
+  obj.shuffle = args.shuffle || args.r || '';
+  obj.condition = args.condition || args.c || {};
+  obj.conditionGraph = args.conditionGraph || {};
+  obj.page = args.page || args.p;
+  if (apiVersion === NcApiVersion.V3 && nested) {
+    if (obj.nestedLimit) {
+      obj.limit = obj.limit = Math.max(
+        Math.min(
+          Math.max(+obj.nestedLimit, 0) || defaultLimitConfig.limitDefault,
+          defaultLimitConfig.limitMax,
+        ),
+        defaultLimitConfig.limitMin,
+      );
+    } else {
+      obj.limit = defaultLimitConfig.ltarV3Limit;
+    }
+  } else {
+    obj.limit = Math.max(
+      Math.min(
+        Math.max(+(args?.limit || args?.l), 0) ||
+          defaultLimitConfig.limitDefault,
+        defaultLimitConfig.limitMax,
+      ),
+      defaultLimitConfig.limitMin,
+    );
+  }
+  obj.offset = Math.max(+(args?.offset || args?.o) || 0, 0);
+  if (obj.page) {
+    obj.offset = (+obj.page - 1) * +obj.limit;
+  }
+  obj.fields =
+    args?.fields || args?.f || (ignoreAssigningWildcardSelect ? null : '*');
+  obj.sort = args?.sort || args?.s || model.primaryKey?.[0]?.column_name;
+  obj.pks = args?.pks;
+  obj.aggregation = args.aggregation || [];
+  obj.column_name = args.column_name;
+  return obj;
+}
+
+export function extractIds(
+  childIds: (string | number | Record<string, any>)[],
+  isBt = false,
+) {
+  return (isBt ? childIds.slice(0, 1) : childIds).map((r) =>
+    typeof r === 'object' ? JSON.stringify(r) : `${r}`,
+  );
+}
+
+export function formatDataForAudit(
+  data: Record<string, unknown>,
+  columns: Column[],
+) {
+  if (!data || typeof data !== 'object') return data;
+  const res = {};
+
+  for (const column of columns) {
+    if (isSystemColumn(column) || isVirtualCol(column)) continue;
+
+    if (!(column.title in data)) {
+      continue;
+    }
+
+    res[column.title] = data[column.title];
+
+    // if multi-select column, convert string to array
+    if (column.uidt === UITypes.MultiSelect) {
+      if (res[column.title] && typeof res[column.title] === 'string') {
+        res[column.title] = (res[column.title] as string).split(',');
+      }
+    }
+    // if attachment then exclude signed url and thumbnail
+    else if (column.uidt === UITypes.Attachment) {
+      if (res[column.title] && Array.isArray(res[column.title])) {
+        try {
+          res[column.title] = (res[column.title] as any[]).map((attachment) =>
+            excludeAttachmentProps(attachment),
+          );
+        } catch {
+          // ignore
+        }
+      }
+    }
+  }
+
+  return res;
 }
